Move use-event into web-vitals package
