--- packages/utilities/use-event/CHANGELOG.md
@@ -1,5 +1,0 @@
-# Use Event Changelog
-
-| Version | Description |
-|---------|-------------|
-| 1.0.0   | [PR#XX](https://github.com/BBC-News/psammead/pull/XX) Initial creation of package. |
