--- packages/utilities/use-event/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@bbc/use-event",
-  "version": "1.0.0",
-  "main": "dist/index.js",
-  "module": "esm/index.js",
-  "sideEffects": false,
-  "description": "Hook for creating single-time event listeners",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/BBC-News/psammead/tree/latest/packages/utilities/use-event"
-  },
-  "author": {
-    "name": "Psammead Maintainers",
-    "email": "PsammeadMaintainers@bbc.co.uk"
-  },
-  "license": "Apache-2.0",
-  "bugs": {
-    "url": "https://github.com/bbc/psammead/issues"
-  },
-  "homepage": "https://github.com/BBC-News/psammead/blob/latest/packages/utilities/use-event/README.md",
-  "dependencies": {},
-  "peerDependencies": {
-    "react": "^16.13.1"
-  }
-}
